fix hash package
