+++ consistenthash/options.go
@@ -1,0 +1,6 @@
+package hash
+
+
+const (
+	minReplicas = 1
+)
